Mod: refactored context providers
